Fix compilation on 64-bit Windows
TBR=pmehta@chromium.org
Review URL: http://codereview.chromium.org/3333015

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@5404 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/spaces.cc
@@ -395,7 +395,7 @@ void MemoryAllocator::FreeRawMemory(void* mem,
 
 void MemoryAllocator::PerformAllocationCallback(ObjectSpace space,
                                                 AllocationAction action,
-                                                int size) {
+                                                size_t size) {
   for (int i = 0; i < memory_allocation_callbacks_.length(); ++i) {
     MemoryAllocationCallbackRegistration registration =
       memory_allocation_callbacks_[i];
